+++ src/FunFair.BuildCheck/ProjectChecks/MustEnableNullable.cs
@@ -12,13 +12,13 @@ namespace FunFair.BuildCheck.ProjectChecks
     [SuppressMessage(category: "ReSharper", checkId: "ClassNeverInstantiated.Global", Justification = "Created by DI")]
     public sealed class MustEnableNullable : IProjectCheck
     {
-        private readonly ILogger<ErrorPolicyWarningAsErrors> _logger;
+        private readonly ILogger<MustEnableNullable> _logger;
 
         /// <summary>
         ///     Constructor.
         /// </summary>
         /// <param name="logger">Logging.</param>
-        public MustEnableNullable(ILogger<ErrorPolicyWarningAsErrors> logger)
+        public MustEnableNullable(ILogger<MustEnableNullable> logger)
         {
             this._logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }