+++ src/FunFair.BuildCheck/ProjectChecks/UsingNSubstituteMustIncludeAnalyzer.cs
@@ -13,7 +13,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         ///     Constructor.
         /// </summary>
         /// <param name="logger">Logging.</param>
-        public UsingNSubstituteMustIncludeAnalyzer(ILogger<NoPreReleaseNuGetPackages> logger)
+        public UsingNSubstituteMustIncludeAnalyzer(ILogger<UsingNSubstituteMustIncludeAnalyzer> logger)
             : base(detectPackageId: @"NSubstitute", mustIncludePackageId: @"NSubstitute.Analyzers.CSharp", logger: logger)
         {
         }