+++ src/FunFair.BuildCheck/ProjectChecks/UsingXUnitMustIncludeTeamCityTestAdapter.cs
@@ -13,7 +13,7 @@ namespace FunFair.BuildCheck.ProjectChecks
         ///     Constructor.
         /// </summary>
         /// <param name="logger">Logging.</param>
-        public UsingXUnitMustIncludeTeamCityTestAdapter(ILogger<NoPreReleaseNuGetPackages> logger)
+        public UsingXUnitMustIncludeTeamCityTestAdapter(ILogger<UsingXUnitMustIncludeTeamCityTestAdapter> logger)
             : base(detectPackageId: @"xunit", mustIncludePackageId: @"TeamCity.VSTest.TestAdapter", logger: logger)
         {
         }